global: configurable job cleanup
* Addresses reanahub/reana#461.

* Removes TTLAfterFinished config since this doesn't work with workflow
  engines because there is RJC as sidecar, and it never completes its execution
  (Flask application). Two solutions were quickly evaluated which either aren't
  too robust or they don't fit into current planning:
  - New endpoint `/shutdown` in RJC, called by RWE when workflow ends
    - The only way found to implement is to explicitly through `RuntimeError`
      inside `/shutdown`, which would make the workflow pod fail.
  - Share folder between containers, write flag to file when workflow finishes.
    - Too convoluted and no time to test thoroughly since we have a workshop
      next week.

+++ reana_workflow_controller/consumer.py
@@ -142,7 +142,8 @@ def _update_workflow_status(workflow, status, logs):
         if status not in ALIVE_STATUSES:
             try:
                 workflow.run_finished_at = datetime.now()
-                _delete_workflow_engine_pod(workflow)
+                if WorkflowStatus.should_cleanup_job(status):
+                    _delete_workflow_engine_pod(workflow)
             except REANAWorkflowControllerError:
                 logging.error(
                     f"Could not clean up workflow engine for workflow {workflow.id_}"
